+++ lib/my_app.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:obs_for_sama/core/constantes.dart';
-import 'package:obs_for_sama/core/routes.dart';
+import 'package:obs_for_sama/core/server_binding.dart';
+import 'package:obs_for_sama/pages/o_b_s_control_page.dart';
+import 'package:obs_for_sama/pages/o_b_s_settings_page.dart';
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
@@ -10,12 +12,24 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return GetMaterialApp(
       key: const ValueKey<String>('Material GetX'),
-      // home: const OBSControlPage(
-      //   key: ValueKey<String>('Main page'),
-      // ),
       theme: kThemeData,
-      onGenerateRoute: Routes().generateRoute,
-      initialRoute: 'home',
+      initialRoute: '/home',
+      getPages: [
+        GetPage(
+          name: '/home',
+          page: () => const OBSControlPage(
+            key: ValueKey<String>('Main page'),
+          ),
+          binding: ServerBinding(),
+        ),
+        GetPage(
+          name: '/settings',
+          page: () => const SettingsPage(
+            key: ValueKey<String>('Settings page'),
+          ),
+          binding: ServerBinding(),
+        ), // here!
+      ],
     );
   }
 }